channel: Make channel id recycling less memory hungry
0e094fe746a9a39f96b5fdc9d8c3874552d4c9d7 introduced reuse of channel id.

The implementation allocates a free list `channels_ids`
holding all **unused** channel ids available.
When the protocol is initialized the list contains 65535 ints, making up
roughly half a MB.

This new implementation instead uses a set() to track **reclaimed**
channel ids. Whenever a new channel id is needed, we first try to take
it from the free set(). If the free set() is empty we forge a new
channel id incrementing a counter.