protocol: Add a handler to the heartbeat, closes: #3
the heartbeat is sent by the server if NO traffic is sent from the server.
so the heartbeat is flushed when any frame is sent, not only heartbeat ones.

the waiter waits for heartbeat * 2

From the documentation:

Heartbeat frames are sent about every timeout / 2 seconds. After two missed heartbeats,
the peer is considered to be unreachable. Different clients manifest this differently but
the TCP connection will be closed. When a client detects that RabbitMQ node is unreachable
due to a heartbeat, it needs to re-connect.