protocol: complete overhaul of heartbeat handling, closes #96
So this patch sets up 2 independent timers.

The first one is the send timer.  This one is triggered after N seconds if
0 bytes have been sent out on the network. Calling write() (or related
methods) on the StreamWriter resets said timer for another N seconds.

The second one is the receive timer.  This one is triggered after N*2
seconds if no traffic has been received.  As soon as bytes are received
from the socket, the timer is reset for another N*2 seconds.