protocol: stop waiting for server to close socket after close_ok frame
The core change is to call _stream_writer.close() both on a
server-initiated close and a client-initiated close, when said close is
ok-ed by the correct side.

This allows us to remove the now-useless stop() method, as
connection_lost() does the same thing (the 'stop_now' future is handled
by run() after the close is effective).

So this allows us to remove the explicit close in tearDown as
AmqpProtocol now handles it properly on its own.