protocol: really turn off heartbeat timers, closes #112
For some reason, calling cancel() on TimerHandle objects returned by
call_later() doesn't really cancel the callback: it just sets the
"cancelled" bit on the handle and the callback will still be called anyway.

Maybe this is a bug that was fixed after 3.4.2, or not, I haven't checked.

But in any case, using None instead of 0 renders both timer callbacks
essentially harmless.