tests: fix flakiness in nack/requeue test
For some reason, this test fails more reliably on python 3.6 than on 3.4.
In any case, the test is wrong, hence this fix.

With the previous code, the delivery callback could be called multiple
times, thus trying to set the future's result many times (which is not
allowed).  Further more, after the first successful delivery, setting the
Future's result would trigger the end of the test, without preventing
further deliveries (since we're nacking the message with a requeue order).

Thus, my hypothesis is that tearDown (which closes() the amqp connection)
would race with the basic_client_nack() (which expects a reply). The
callback would thus block AmqpProtocol's worker Task which dispatches
frames received from the server.

2 take aways:

1. properly finish the test (in this case, stop further delivery) before
   closing the connection to prevent a race
2. calling back into amqp from a consume callback is... messy. We should
   avoid it (though we're not in this case)