channel: make publish() and basic_publish() use _write_frame()
There is a danger: if different tasks try to use methods from the same
channel at the same time, there is a risk that publish() and
basic_publish() frames get intertwined with unrelated frames, wreaking
havoc.

But right now, this is impossible because _write_frame isn't a real
coroutine that yields to the main loop.  It's simply a @coroutine-decorated
function.

Also, remove the 'is_open' check because _write_frame does it too.